Solved some problems exposed by clang-tidy

+++ src/game/update/play/sequence/internal.h
@@ -17,7 +17,7 @@ int game_update_play_sequence_get_hallway_height(
     const struct game_state_play *self);
 int game_update_play_sequence_get_gap(const struct game_player *self);
 float game_update_play_sequence_get_width(const struct game_player *self,
-    float gap);
+    int gap);
 enum game_sequence_object_type game_update_play_sequence_get_next_object_type(
     size_t current_index, const struct game *game);
 enum game_sequence_object_type game_update_play_sequence_get_one_of_first_two(
